parallelize snp_readBed2()
