+++ man/snp_readBed.Rd
@@ -11,7 +11,8 @@ snp_readBed2(
   bedfile,
   backingfile = sub_bed(bedfile),
   ind.row = rows_along(obj.bed),
-  ind.col = cols_along(obj.bed)
+  ind.col = cols_along(obj.bed),
+  ncores = 1
 )
 }
 \arguments{
@@ -29,6 +30,9 @@ are used. If not specified, all rows are used.\cr
 \item{ind.col}{An optional vector of the column indices (SNPs) that are used.
 If not specified, all columns are used.\cr
 \strong{Don't use negative indices.}}
+
+\item{ncores}{Number of cores used. Default doesn't use parallelism.
+You may use \link{nb_cores}.}
 }
 \value{
 The path to the RDS file that stores the \code{bigSNP} object.
