+++ tests/testthat/test-1-readBed.R
@@ -69,7 +69,7 @@ test_that("snp_readBed2() works", {
   ind.col <- sample(ncol(G), ncol(G) / 2, replace = TRUE)
 
   test2 <- snp_attach(snp_readBed2(bedfile, backingfile = tempfile(),
-                                   ind.row, ind.col))
+                                   ind.row, ind.col, ncores = 2))
   test3 <- snp_attach(subset(test, ind.row, ind.col))
   expect_equal(test2$genotypes[], test3$genotypes[])
   expect_equal(test2[-1], test3[-1])
